fix: Leave keys as buffers in example.mjs & support b4a buffer in apply
The key in `simplePut(key)` was cast to string because of an encoding
issue in `Autobee.apply`. The `view.put(key)` didn't accept the json
decoded value as a valid buffer because it was of the following form
after decoding:

```json
{
  type: 'Buffer',
  data: [
     65, 186,  39,  76, 106, 245, 134,
    239, 106, 105, 228,  51,  50,  38,
    234, 252, 183,  55, 228, 119, 241,
    217, 134, 195, 188, 111, 102,  10,
    224,  30, 196, 133
  ]
}
```

Using `b4a.from()` allows support for this encoding of a buffer as well.